pyproject.toml
[project]
name = "entity-extraction-pipeline-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apache-airflow==2.10.5",
    "apache-airflow-providers-docker>=4.3.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5", "ruff>=0.11.6"]
